Fix curl fetcher response / status code reporting
The full output of the curl command was used to record results in the
histogram. Use just the HTTP status code instead.

There can still be cases when a result other than an HTTP status code
is recorded, such as if curl prints a "Could not resolve host" error.
This seems fine, and informative to the user.

Also follow redirects. This is the default behavior of the Go HTTP
client, and therefore the go fetcher in hammer. Now the curl fetcher
does the same.